shadow creation and tests
* add (some) RefObject.Get/SetValue support for pointers and array of
pointers
* bug fix ops/shadow by adding Addr() to the target / spec writer
interface
* bug fix shadow class GetObject: return the constructed object on
success.
* let more spec assignments fall into the default channel in order to
handle a wider variety of underlying number types. ( ex. state enums
which are numbers, but not explicitly so )
* rename ops’s targetPtr to target because hasnt actually been a
pointer since the shadow introduction; rename _SpecWriter to _Target to
give a better hint as to what it is.
* implement shadow evalFromType — its very similar to literally, but
requires returning an eval interface and not a eval implementation
* re-implement ops test using the existing go lang testing.T.Run (
instead of testify.suite )
* disable std test till it works