composer: change tool config to use go install directory
go build ./... doesnt save the cmds, and there doesnt appear to be a way to force it to.
the only real option to rebuild all tools seems to be go install ./... instead.
why the golang team chose to do this, and why the behavior is so poorly documented, is a mystery.